+++ ontimize/config/o-providers.ts
@@ -4,9 +4,30 @@ import { BaseRequestOptions, XHRBackend } from '@angular/http';
 import { MAT_RIPPLE_GLOBAL_OPTIONS } from '@angular/material';
 import { Router } from '@angular/router';
 import { combineLatest } from 'rxjs';
+
 import { OContextMenuService } from '../components/contextmenu/o-context-menu.service';
 import { AppConfig, Config } from '../config/app-config';
-import { appConfigFactory, AppMenuService, AuthGuardService, CurrencyService, dataServiceFactory, DialogService, LocalStorageService, LoginService, MomentService, NavigationService, NumberService, OModulesInfoService, OntimizeExportService, OntimizeFileService, OntimizeMatIconRegistry, OntimizeService, OntimizeServiceResponseParser, OTranslateService, OUserInfoService, SnackBarService } from '../services';
+import {
+  appConfigFactory,
+  AppMenuService,
+  AuthGuardService,
+  CurrencyService,
+  dataServiceFactory,
+  DialogService,
+  LocalStorageService,
+  LoginService,
+  MomentService,
+  NavigationService,
+  NumberService,
+  OModulesInfoService,
+  OntimizeFileService,
+  OntimizeMatIconRegistry,
+  OntimizeService,
+  OntimizeServiceResponseParser,
+  OTranslateService,
+  OUserInfoService,
+  SnackBarService
+} from '../services';
 import { OFormLayoutManagerService } from '../services/o-form-layout-manager.service';
 import { Error403Component } from '../services/permissions/error403/o-error-403.component';
 import { ORemoteConfigurationService } from '../services/remote-config.service';
@@ -29,23 +50,34 @@ export function appInitializerFactory(injector: Injector, config: Config, oTrans
     let observableArray = [];
     const locationInitialized = injector.get(LOCATION_INITIALIZED, Promise.resolve(null));
     locationInitialized.then(() => {
-      oTranslate.setDefaultLang('en');
-      let userLang = config['locale'];
-      if (!userLang) {
-        // use navigator lang if available
-        userLang = oTranslate.getBrowserLang();
+      const storedLang = oTranslate.getStoredLanguage();
+      const configLang = config['locale'];
+      const browserLang = oTranslate.getBrowserLang();
+      let userLang = 'en';
+      let defaultLang = 'en';
+      if (storedLang) {
+        userLang = storedLang;
+      } else if (configLang) {
+        userLang = configLang;
+        defaultLang = configLang;
+      } else if (browserLang) {
+        userLang = browserLang;
+        defaultLang = browserLang;
       }
+      oTranslate.setDefaultLang(defaultLang);
+
+      const locales = new Set(config.applicationLocales || []);
+      locales.add('en');
+      locales.add(userLang);
+
+      const localseArray = [];
+      locales.forEach((val1, val2, self) => localseArray.push(val1));
 
       // initialize available locales array if needed
       if (!config.applicationLocales) {
-        config.applicationLocales = [];
+        config.applicationLocales = localseArray; // use [...locales] with es6
       }
-      if (config.applicationLocales.indexOf('en') === -1) {
-        config.applicationLocales.push('en');
-      }
-      if (userLang && config.applicationLocales.indexOf(userLang) === -1) {
-        config.applicationLocales.push(userLang);
-      }
+
       if (config['uuid'] === undefined || config['uuid'] === null || config['uuid'] === '') {
         console.error('Your app must have an \'uuid\' property defined on your app.config file. Otherwise, your application will not work correctly.');
         alert('Your app must have an \'uuid\' property defined on your app.config file. Otherwise, your application will not work correctly.');
@@ -98,10 +130,6 @@ export function getOntimizeServiceProvider(backend: XHRBackend, defaultOptions: 
 
 export function getOntimizeFileServiceProvider(injector: Injector) {
   return new OntimizeFileService(injector);
-}
-
-export function getOntimizeExportServiceProvider(injector: Injector) {
-  return new OntimizeExportService(injector);
 }
 
 export function getLoginServiceProvider(injector: Injector) {
@@ -190,11 +218,6 @@ export const ONTIMIZE_PROVIDERS: Provider[] = [
   {
     provide: OntimizeFileService,
     useFactory: getOntimizeFileServiceProvider,
-    deps: [Injector]
-  },
-  {
-    provide: OntimizeExportService,
-    useFactory: getOntimizeExportServiceProvider,
     deps: [Injector]
   },
   // getLoginServiceProvider