+++ ontimize/types.ts
@@ -1,2 +1,3 @@
+export * from './types/export-service.interface';
 export * from './types/ontimize-service-config.type';
 export * from './types/remote-configuration.type';